Fix build with Ruby 2
test/unit in Ruby 2 is doing all sorts of fancy shit if stdout is a tty,
which messes up the test. So just use pipes for stdout and stderr. stdin
still needs to be a pty otherwise some output will end up in the
terminal where the spring tests are being run.

See also #58