Avoid using pty in tests
It was causing too many issues and was only being used in our tests
anyway, see #58.

Note that this means that I've had to remove the error if stdin is not a
tty, but I don't really think that matters much. I'm not sure why anyone
would use spring in a situation where stdin is not a tty.