Stop preloading spec_helper and test_helper by default
This causes subtle issues, such as #113.

I did this because I find it useful to preload my spec_helper in the
project I work on (because it is slow to require, since it refreshes the
database schema).

However, I have decided that the behaviour is too surprising to be a
default. It's more important that Spring works well out of the box than
it is to eke out every last ounce of performance in every situation.
Therefore, I've suggested preloading these files in the README, but we
will no longer do it by default.