Use remove_connection rather than connection.disconnect!
remove_connection will work even if we haven't actually connected to the
database yet. But connection.disconnect! will cause a connection to be
made and then disconnected - and this will fail if we can't create a
connection (e.g. because we're running rake db:create and the database
doesn't exist yet).

Fixes #125.