Don't detect application root in the client
This requires reading config/spring.rb, which we don't want to do
because it may depend on additional gems.

Instead, just boot the server. If the application root can't be found,
the server will throw an error.

To support this I added some extra testing stuff in order to capture the
output that the server writes to the TTY. (Previously it was just going
ot /dev/null in the tests.)

This will also be helpful for debugging issues on the CI.

Fixes #144.