Don't delete bundler environment variables
We need to snapshot the ENV at the earliest possible moment, i.e. right
at the top of lib/spring/server.rb. And we need to not delete those env
vars that were changed.

For example: suppose I have export RUBYOPT="-rubygems" in my .bashrc.

When I first run spring, bundler will add to RUBYOPT so that it includes an
"-rbundler/setup".

When I next run spring, the client will send RUBYOPT="-rubygems" to the
server, but the server shouldn't just use that value, because the value
go changed when the server started, so we can assume that the process of
starting the application causes that variable to be changed, and
therefore we should stick with the server's version.

Fixes #205.