Avoid running the reloaders when no files have changed
Rails now reloads the routes.rb on every invocation of AD::Reloader (see
https://github.com/rails/rails/commit/b9b06daa915fdc4d11e8cfe11a7175e5cd8f104f).
This degrades the performance of running a spring command, at least in
my application which has a large routes file. This change doesn't
improve the performance, but it avoids running the reloader when we
don't need to.

Closes #211.

This is a partial fix. See #212 for details about how this can be
improved.