Avoid FileUtils
I'm getting sporadic failures on Travis CI due to this part of the code.

What *seems* to be happening is that the FileUtils.rm_rf call is either
not properly removing everything, or it's returning before everything
has actually been deleting.

This means that when we run the cp_r, path already exists, so instead of
application.root being copied *as* path, it's copied *into* path.

Shelling out should hopefully be more reliable.