Ensure the spring binfile from the correct gem is loaded
It turns out that Gem.bin_path won't return the bin path from the
already-activated spring gem if there is a newer version on the system -
in such a case it'll return the newer one, so we'll end up loading
lib/spring/binstub.rb from the old version, which will then load
bin/spring from the newer version.

So we shouldn't rely on Gem.bin_path here, as we want to load the
bin/spring from the version of the gem which has already been activated.

Fixes #238.