Fix usage of uninitialized ::DL in Ruby 1.9.3
When compiling Ruby 1.9.3 with clang, Fiddle is compiled in while DL is not.
This commit fixes an usage of the DL module in the case it's not compiled in.

Unfortunately Fiddle::Handle and Fiddle.dlopen were only added in Ruby 2, but
a dlopen call (available as part of the DL module) is needed to use
Fiddle::Function.

Fixes #274