Extract listen watcher to separate gem
https://github.com/jonleighton/spring-watcher-listen

The reason to do this is that I want to minimise the amount of code
which needs to be maintained in the core spring project. I don't use the
listen watcher and therefore don't want to maintain it. If it's a
separate gem, others can maintain it on their own schedule.