Fix a problem with gems bundled from a git repository
This occurred when the `bin/spring` was generated before 1.3.0.

Upgrading to 1.3.0 then lead to a bundler error. This happened because
in 1.3.0 we switched from forking the server process to spawning it as a
fresh process. In the fresh process, bundler didn't seem to like
GEM_HOME being set to an empty string.

We have already separately changed the generated binstub so that
GEM_HOME is nil, rather than the empty string
(666db7ddbdf990f9ef03cc52c872c4e402cc543d). Therefore this fix just maps
the empty string from older binstubs to a nil value.