Fix setting of GEM_HOME
In 1.3.0 we started to launch the Spring server process via a
Process.spawn, rather than a fork
(ae175de3dd6dbae6d9aef59dbbfeed5a5eec90df).

This brought to light some problems with how we were handling the
GEM_HOME environment variable. In < 1.3 it was set to an empty string.
In 1.3.0 it was set to nil.

The bug was reported in #352 and #383.

After switching to Process.spawn, having no value in GEM_HOME in the
server process meant that Bundler was unable to find bundled git
repositories.

In actual fact, we do not need to set the gem home in the binstub. We
only need to set the gem path, so that the install of spring which is in
the bundle can be found. And we don't really need to mutate the ENV hash
either.

This makes the binstub clearer and simpler.

We then set the GEM_PATH and GEM_HOME as we need them for the env when
we spawn the server process.