Base the pidfile path on the socket path
This will improve support for running spring inside a Docker container,
since you can set the SPRING_SOCKET to be at a different path
inside/outside the container, and the pidfile path will automatically
match that. This if the pid namespace is shared, setting SPRING_SOCKET
appropriately inside and outside of the container will enable "spring
status" and "spring stop" to work properly (since these commands rely on
being able to find the pidfile and associated process).